fix audio kill bug for macos

+++ lab4_2/lib/audio.cpp
@@ -103,14 +103,17 @@ void Audio::play_loop() {
 }
 
 void Audio::stop() {
-    _loop = false;
-    int status = kill(this->_loop_pid, 0);
-    if (status != -1) {
-        kill(this->_loop_pid, SIGKILL);
+    if (_loop_pid != 0) {
+        _loop = false;
+        int status = kill(this->_loop_pid, 0);
+        if (status != -1) {
+            kill(this->_loop_pid, SIGKILL);
+        }
+        _t->join();
+        delete _t;
+        _t = nullptr;
+        _loop_pid = 0;
     }
-    _t->join();
-    delete _t;
-    _t = nullptr;
 }
 
 #endif
